fix: ArticleWithCount querydsl 사용시 타입이 안맞는부분 수정

+++ src/main/kotlin/com/sat/board/infrastructure/ArticleRepositoryImpl.kt
@@ -25,8 +25,8 @@ class ArticleRepositoryImpl(
                 `as`(article.id, "id"),
                 `as`(article.title, "title"),
                 `as`(category.name.value, "category"),
-                `as`(comment.countDistinct(), "commentCount"),
-                `as`(like.countDistinct(), "likeCount"),
+                `as`(comment.countDistinct().intValue(), "commentCount"),
+                `as`(like.countDistinct().intValue(), "likeCount"),
                 `as`(article.createdDateTime, "createdDateTime"),
             ))
             .from(article)
